+++ src/main/java/racingcar/view/InputView.java
@@ -1,9 +1,9 @@
-package racingcar;
+package racingcar.view;
 
 import java.util.Scanner;
 
 public class InputView {
-    
+
     private static Scanner scanner = new Scanner(System.in);
     private static final int MINIMUM_TRY_COUNT = 1;
