+++ src/main/java/racingcar/view/OutputView.java
@@ -1,7 +1,9 @@
-package racingcar;
+package racingcar.view;
 
 import java.util.List;
 import java.util.stream.Collectors;
+import racingcar.domain.Car;
+import racingcar.domain.Cars;
 
 public class OutputView {
