+++ src/test/java/racingcar/RacingcarServiceTest.java
@@ -9,6 +9,8 @@ import java.util.Arrays;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
+import racingcar.domain.Cars;
 
 class RacingcarServiceTest {
